depuracion de errores

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: mmiro-go <mmiro-go@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/12 18:11:55 by mmiro-go          #+#    #+#             */
-/*   Updated: 2023/01/26 15:23:01 by mmiro-go         ###   ########.fr       */
+/*   Updated: 2023/02/09 14:33:27 by mmiro-go         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i++ < n)
